chore: clean up tokenHandling slice
We've also updated the version of user-menu-component.
Since the id token claims are already made available in the login slice
(as userProfile), there is no need to also expose them in the
tokenHandling slice.
The new version of the user-menu-component adds better support for OIDC
logout which requires the raw value of the id token. Thus we replace the
value in id token with the actual id token.